Move MutableChain to scrapy.utils.python
